requirements_old.txt
annotated-types==0.6.0
anyio==4.3.0
click==8.1.7
exceptiongroup==1.2.0
fastapi==0.99.1
h11==0.14.0
httptools==0.6.1
idna==3.6
pydantic_core==2.16.3
python-dotenv==1.0.1
PyYAML==6.0.1
sniffio==1.3.1
typing_extensions==4.10.0
uvicorn==0.29.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
fastapi-jwt-auth==0.5.0
psycopg2-binary==2.9.9
SQLAlchemy==2.0.28